feat: Ollama serve harness + v2 GGUF format + Qwen3 Q2_0 importer
Three features landing together:

1. Ollama-compat HTTP server ("serve" subcommand):
   - Ollama-native /api/{version,tags,show,chat,generate,embeddings}
   - OpenAI-compat /v1/{models,chat/completions,completions}
   - NDJSON streaming for Ollama, SSE for OpenAI, correct done/[DONE] terminators
   - Snake_case JSON, tag-suffix tolerance (:latest + custom tags)
   - Backed by existing BitNetHostedAgentModel; no model changes needed
   - 45 route-shape + framer + registry tests green

2. v2 GGUF format (ggml_type 1001):
   - Packs BitNet ternary weights as trits + per-tensor FP32 Gamma
   - 19.95x shrink vs v1 on real 8B Bonsai (25.88 GB -> 1.30 GB)
   - Denser than source Q2_0 (5 trits/byte base-3 beats 2 bits + FP16)
   - GgufStreamingReader/Writer for large-file memory discipline

3. Qwen3 Bonsai Q2_0 importer ("import-gguf" subcommand):
   - PrismQ2_0 decoder (34-byte block, quaternary -> ternary collapse)
   - Analytic Gamma preserves E[|w|] across the lossy collapse
   - GroupedQueryAttention (32Q/8KV) + Qwen3Like8B config preset
   - BitLinear.ImportTernary direct integer path (no FP32 detour)
   - Discards tokenizer-mismatched token_embd + output, re-seeds
   - Separate BitNetSharp.Converter.Tests assembly (52 tests)

Test counts: 500 Core + 52 Converter + 4 Runtime = 556 green.

+++ src/BitNetSharp.Core/BitNetVisualizer.cs
@@ -31,7 +31,7 @@ public sealed class BitNetVisualizer
     private static string RenderWeightHistogram(TrainingReport trainingReport)
     {
         var max = new[] { trainingReport.NegativeWeights, trainingReport.ZeroWeights, trainingReport.PositiveWeights }.Max();
-        max = Math.Max(max, 1);
+        max = Math.Max(max, 1L);
 
         return string.Join(
             Environment.NewLine,
@@ -57,7 +57,7 @@ public sealed class BitNetVisualizer
         return builder.ToString();
     }
 
-    private static string FormatBar(string label, int value, int max)
+    private static string FormatBar(string label, long value, long max)
     {
         var width = Math.Max(1, (int)Math.Round(value / (double)max * 20d));
         return $"{label}: {new string('#', width)} {value}";
